Apply rebuild tag ranges
These tag ranges are more informative than the poorly defined "mass-rebuild" tags.

Apologies for the repetitive code. I haven't written PHP in years and just wanted to bang this out quickly over the GitHub editor (all I have right now), so I wanted to keep it as stupid as possible to minimize the chance of screwups on untested code.